+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Switch Units option from temperature only (celsius, fahrenheit) to system (metric, imperial)
 
+### Fixed
+
+- Cache missing language and unit
+
 ## [0.10.0] - 2023-03-23
 
 ### Added